requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cohere==5.13.0
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
executing==2.2.0
fastapi==0.116.1
fastavro==1.12.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.8
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-cohere==0.4.4
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.17
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.101.0
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
parameterized==0.9.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
postgrest==1.1.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.2
realtime==2.7.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tavily-python==0.7.11
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-PyYAML==6.0.12.20250822
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
